Worker/requirements.txt
dnspython==2.7.0
pika==1.3.2
psycopg2-binary==2.9.10
pymongo==4.10.1
python-dotenv==1.0.1
secure-smtplib==0.1.1
